building new helper function to handle user validation with hooks
